+++ api/src/main/java/com/techs/api/model/UserDto.java
@@ -1,0 +1,9 @@
+package com.techs.api.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record UserDto(
+    @JsonProperty("id") String id,
+    @JsonProperty("name") String name,
+    @JsonProperty("email") String email
+) {}